Fixes syntax errors. Function moved into existing magic module.
