+++ client/homebrew/editor/snippetbar/snippets/snippets.js
@@ -6,7 +6,6 @@ const MonsterBlockGen = require('./monsterblock.gen.js');
 const ClassFeatureGen = require('./classfeature.gen.js');
 const CoverPageGen = require('./coverpage.gen.js');
 const TableOfContentsGen = require('./tableOfContents.gen.js');
-const MagicItemGen = require('./magicitem.gen.js');
 
 
 module.exports = [
@@ -147,7 +146,7 @@ module.exports = [
 			{
 				name : 'Magic Item',
 				icon : 'fa-key',
-				gen  : MagicItemGen,
+				gen  : MagicGen.item,
 			},
 		]
 	},
